pyproject.toml
[project]
name = "jobscraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.0",
    "python-jobspy>=1.1.82",
    "requests>=2.32.0",
]
